TodayOnHistory/requirements.txt
flask
pillow
bs4
requests